[frontend] resolve PR comments (#issue/3242)
