+++ opencti-platform/opencti-front/src/components/ExportButtons.jsx
@@ -12,7 +12,7 @@ import Tooltip from '@mui/material/Tooltip';
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
 import ToggleButton from '@mui/material/ToggleButton';
 import { withRouter } from 'react-router-dom';
-import fileDownload from 'js-file-download';
+import handleExportJson from '../private/components/workspaces/workspaceExportHandler';
 import themeLight from './ThemeLight';
 import themeDark from './ThemeDark';
 import { commitLocalUpdate } from '../relay/environment';
@@ -109,24 +109,6 @@ class ExportButtons extends Component {
     this.setState({ anchorElPdf: event.currentTarget });
   }
 
-  handleExportJson(workspace) {
-    this.setState({ exporting: true });
-
-    const dashboardName = workspace.name;
-    const dashboardConfig = JSON.stringify({
-      version: '1.0.0',
-      type: workspace.type,
-      name: dashboardName,
-      manifest: workspace.manifest,
-    }, null, 2);
-    const blob = new Blob([dashboardConfig], { type: 'text/json' });
-    const [day, month, year] = new Date().toLocaleDateString('fr-FR').split('/');
-    const fileName = `${year}${month}${day}_octi_dashboard_${dashboardName}`;
-
-    fileDownload(blob, fileName, 'application/json');
-    this.setState({ exporting: false });
-  }
-
   handleClosePdf() {
     this.setState({ anchorElPdf: null });
   }
@@ -206,7 +188,7 @@ class ExportButtons extends Component {
           </Tooltip>
           {isCustomDashBoard && (
             <Tooltip title={t('Export to JSON')}>
-              <ToggleButton onClick={this.handleExportJson.bind(this, workspace)}>
+              <ToggleButton onClick={() => handleExportJson(workspace)} value={'Export-to-JSON'}>
                 <FileExportOutline fontSize="small" color="primary" />
               </ToggleButton>
             </Tooltip>
